Правки по модуляции
