+++ js/create-gallery-of-mock-data.js
@@ -1,33 +1,8 @@
 import {getRandomInteger} from './util.js';
 import {createRandomNumberFromRangeGenerator} from './util.js';
+import {getRandomArrayElement} from './util.js';
+import {NAMES, DESCRIPTION_LIST, MESSAGE_LIST} from './data.js';
 
-const NAMES = [
-  'Артём',
-  'Екатерина',
-  'Саша',
-  'Юлия',
-  'Иван',
-  'Мария',
-  'Виктор'
-];
-
-const DESCRIPTION_LIST = [
-  'Ещё один счастливый день.',
-  'Жду лайки!',
-  'Помотрите, какая милота.',
-  'Комментарии здесь неуместны, просто любуемся.'
-];
-
-const MESSAGE_LIST = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 const getRandomCommentId = createRandomNumberFromRangeGenerator(0, 999);
 const getRandomAvatarIndex = createRandomNumberFromRangeGenerator(1, 6);
 const getRandomPhotoId = createRandomNumberFromRangeGenerator(1, 25);
