+++ js/util.js
@@ -24,5 +24,8 @@ const createRandomNumberFromRangeGenerator = (min, max) => {
   };
 };
 
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
 export {getRandomInteger};
 export {createRandomNumberFromRangeGenerator};
+export {getRandomArrayElement};
